pdf2text/requirements.txt
requests
pdf2images
pytesseract
opencv-python